requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
constantly==15.1.0
cryptography==41.0.4
daphne==4.0.0
Django==5.0.6
django-admin-honeypot==1.1.0
django-admin-honeypot-updated-2021==1.2.0
django-admin-thumbnails==0.2.8
django-ckeditor-5==0.2.13
django-debug-toolbar==4.2.0
django-ipware==5.0.0
django-js-asset==2.1.0
django-mptt==0.15.0
django-storages==1.13.2
django-translations==1.3.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
midtransclient==1.4.2
msgpack==1.0.7
Pillow==10.0.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
python-decouple==3.8
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
service-identity==23.1.0
setuptools==70.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
urllib3==2.0.4
zope.interface==6.0
